+++ src/main/java/com/smartshop/shop/dto/responseDTO/OrderItemResponseDTO.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,6 +11,6 @@ public class OrderItemResponseDTO {
     private String id;
     private String productName;
     private Integer quantity;
-    private BigDecimal prixUnitaire;
-    private BigDecimal totalLigne;
+    private double prixUnitaire;
+    private double totalLigne;
 }
