+++ src/main/java/com/smartshop/shop/dto/responseDTO/PaymentResponseDTO.java
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -15,7 +14,7 @@ import java.time.LocalDate;
 public class PaymentResponseDTO {
     private String id;
     private Integer paymentNumber;
-    private BigDecimal montant;
+    private double montant;
     private LocalDate datePayment;
     private PaymentStatus status;
     private TypePayment typePayment;
