+++ src/main/java/com/smartshop/shop/dto/responseDTO/ProductResponseDTO.java
@@ -4,15 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductResponseDTO {
     private String id;
     private String nom;
-    private BigDecimal prixUnitaire;
+    private double prixUnitaire;
     private Integer stockDisponible;
     private Boolean deleted;
 }
